Add ability to resume uploads and downloads
In a streaming situation, where the full file content is not available,
it is not always possible to provide more data when the source callback
is invoked.  This change allows the API user to resume uploading data
once more has streamed in.

In the download case, the receiver may not yet be ready to handle more
data, and so now it may be held until it can be consumed.

Resolves: LE-12351
Change-Id: I94d25f1d7a5101d23d06d5cf779519ef1ccf1964