[HLLEAPP-609] Duplicate KFTP_ERROR URC returned when delete file
When deleting a file using kftpdel command when the file is
unavailable, duplicate KFTP_ERROR URC is returned. This is because the
response code is not cleared when the timer times out on the client
side. The solution is to clear the response code when the timer times
out from the client side.

Resolves: HLLEAPP-609
Change-Id: If5869922189d742f6e395982a83c7dfb3d9161b0
Signed-off-by: Matthew Law <mlaw@sierrawireless.com>