[HLLEAPP-615] Timers are active during transfer
Both control and data timers are active during data transfer. The
timers should be stopped or reset during FTP transfers.
The solution is to stop the control timer after receiving the response
code and to reset the data timer when the device receives more data.
This is the same situation for FTP send. When sending a large FTP file,
the control timer will timeout and will no longer transfer to the FTP
server.

Change-Id: Ib48d83c828b27393120a8fb39511ca62ea05bc31
Resolves: HLLEAPP-615
Signed-off-by: Matthew Law <mlaw@sierrawireless.com>