[LE-12989] Unable to start send or receive FTP data after running
Issue Desription:
After running AT+KFTPSND the first time, you will not be able to run
anything else unless you disconnect from the FTP server and reconnect
again.

Root Cause:
LWFTP state is stuck in XFERING and cannot be switched back to LOGGED
unless it is disconnected from the FTP server and reconnected again.

Solution:
Fix the state so that it does not get stuck in XFERING state after
initial transfer.
Also fixed an issue where if KFTPRCV ran the second time can only
transfer limited amount of data. This is because the recv_done flag is
not set to 0 at the beginning of every KFTPRCV.

Notes:
Added #ifndef NO_SWI so that it shows original file.

Resolves: LE-12989
Change-Id: I291b4bc9aba34133314eef9674295f92327cc100
Signed-off-by: Matthew Law <mlaw@sierrawireless.com>