[LE-13903] AT port is locked after sending kftprcv
Issue Description:
AT port does not respond after doing kftprcv

Root cause:
The receive done callback will never be triggered when sending small
files because of the previous change. LE-13835

Solution:
Cannot merge the if statements. Must use two if statements so that it
will detect when the buffer is empty and trigger the receive done
callback.

Resolves: LE-13903
Change-Id: Iedb7710148ed16aa5d96a46ee5e164dda22d6554
Signed-off-by: Matthew Law <mlaw@sierrawireless.com>