[HLLEAPP-775] Memory leak hangs module if FTP response is spread over multiple packets
Problem:
If an FTP response code was sent over multiple packets, the
previous packet would be invalidated and discarded. When being
discarded, however, the pbuf would not be freed resulting in a memory
leak.

If this continued over enough FTP responses, the pbuf pool would become
empty and the module would no longer be able to receive any TCP
acknowledgements.

Additionally, it was found that the pbuf payload did not contain a null
terminator. This occasionally resulted strtok_r() returning a garbage
string if adjacent memory was non-0.

Although I did not ever see this cause an error during testing, it is
conceivable that the garbage string start with numbers and therefore be
interpreted as an incorrect response code.

Solution:
The pbuf is now freed upon being invalidated.

Additionally, the pbuf payload is now first copied to a buffer with
space for a null terminator.

Resolves: HLLEAPP-775
Related: EURY-3591
Change-Id: Ib5a3d34cf6d16be2ff20fdcc87c6daa468ba3f57
Signed-off-by: Grayson Halliday <Ghalliday@sierrawireless.com>