[LE-15522] AT+KFTPLS returns 0 file size
Issue:
When lwftp receives an FTP response, the original data buffer
does not contain a null terminator. Given string operations are to be
performed on the data, the payload is first copied into a separate
buffer in which a null terminator is added.

This additional buffer, however, was being freed too early. The SIZE
operation handling was unique in that it referenced the FTP response
text which was previously being freed after parsing for the response
code.

Solution:
The additional buffer is now freed after FTP response handling.

Resolves: LE-15522
Related: EURY-4059
Change-Id: I8c3925a3b89510f4ee22aed55c82eee031881c6c
Signed-off-by: Grayson Halliday <GHalliday@sierrawireless.com>